Move general store to base
