+++ src/app/base/base.module.ts
@@ -3,9 +3,23 @@ import { NgModule } from '@angular/core';
 import { FlexLayoutModule } from '@ngbracket/ngx-layout';
 import { ReactiveFormsModule } from '@angular/forms';
 
+import { StoreModule } from '@ngrx/store';
+import { EffectsModule } from '@ngrx/effects';
+import { reducer, featureName, CuttingSheetsEffects } from './store';
+
 @NgModule({
   declarations: [],
-  imports: [CommonModule, FlexLayoutModule, ReactiveFormsModule],
-  exports: [CommonModule, ReactiveFormsModule, FlexLayoutModule],
+  imports: [
+    StoreModule.forFeature(featureName, reducer),
+    EffectsModule.forFeature([CuttingSheetsEffects]),
+    CommonModule,
+    FlexLayoutModule,
+    ReactiveFormsModule
+  ],
+  exports: [
+    CommonModule,
+    ReactiveFormsModule,
+    FlexLayoutModule
+  ],
 })
 export class BaseModule {}
