+++ src/app/base/guards/index.ts
@@ -1,0 +1,2 @@
+export { JobTypesGuard } from './job-types.guard';
+export { StageTemplatesGuard } from './stage-templates.guard';
