+++ src/app/base/guards/job-types.guard.ts
@@ -3,7 +3,7 @@ import { UrlTree } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 
-import { loadEntity } from '../../../core';
+import { loadEntity } from '../../core';
 import * as fromFeature from '../store';
 
 @Injectable({ providedIn: 'root' })
