+++ src/app/base/models/index.ts
@@ -1,0 +1,2 @@
+export { StageTemplate, StageTemplateServer } from './stage-template';
+export { Type } from './type';
