+++ src/app/base/models/stage-template.ts
