+++ src/app/base/services/index.ts
@@ -1,0 +1,2 @@
+export { JobTypesService } from './job-types.service';
+export { StageTemplatesService } from './stage-templates.service';
