+++ src/app/base/services/stage-templates.service.ts
