+++ src/app/base/store/actions/stage-templates.actions.ts
