+++ src/app/base/store/index.ts
@@ -1,0 +1,4 @@
+export { CuttingSheetsEffects } from './effects/effects';
+export * from './actions';
+export * from './reducers';
+export * from './selectors';
