+++ src/app/base/store/reducers/job-types.reducer.ts
