+++ src/app/base/store/reducers/stage-templates.reducer.ts
