+++ src/app/base/store/selectors/stage-templates.selectors.ts
