+++ src/app/features/cutting-sheets/models/index.ts
@@ -1,3 +1,3 @@
 export * from './cutting-sheet';
 export * from './stage';
-export * from './type';
+export * from '../../../base/models/type';
