--- src/app/features/cutting-sheets/store/effects/index.ts
@@ -1,1 +1,0 @@
-export { CuttingSheetsEffects } from './effects';
