--- src/app/features/cutting-sheets/store/index.ts
@@ -1,4 +1,0 @@
-export * from './actions';
-export * from './effects';
-export * from './reducers';
-export * from './selectors';
